Remove unused DOM elements
Changes effectively nothing. Just saves a few unnecessary bytes.
One of the removed variables was the DOM element for the retro computer monitor frame (var overlay = document.getElementById("overlay")), which ended up having no Javascript used on it, and the other was for the ASCII art head at the top (var docHead = document.getElementById("head")), which also never had JS used on it.